INS-568 split pulsar to different files. moving consensus methods to pulsarconsensus. moving timers to pulsartimers
