+++ pulsar/pulsarhelpers.go
@@ -26,21 +26,88 @@ import (
 
 	ecdsahelper "github.com/insolar/insolar/cryptohelpers/ecdsa"
 	"github.com/insolar/insolar/cryptohelpers/hash"
+	"github.com/insolar/insolar/log"
 	"github.com/pkg/errors"
 	"github.com/ugorji/go/codec"
 
 	"github.com/insolar/insolar/core"
 )
 
+// FetchNeighbour searches neighbour of the pulsar by pubKey of a neighbout
+func (currentPulsar *Pulsar) FetchNeighbour(pubKey string) (*Neighbour, error) {
+	neighbour, ok := currentPulsar.Neighbours[pubKey]
+	if !ok {
+		return nil, errors.New("forbidden connection")
+	}
+	return neighbour, nil
+}
+
+// IsStateFailed checks if state of the pulsar is failed or not
+func (currentPulsar *Pulsar) IsStateFailed() bool {
+	return currentPulsar.StateSwitcher.GetState() == Failed
+}
+
+func (currentPulsar *Pulsar) getMaxTraitorsCount() int {
+	nodes := len(currentPulsar.Neighbours) + 1
+	return (nodes - 1) / 3
+}
+
+func (currentPulsar *Pulsar) getMinimumNonTraitorsCount() int {
+	nodes := len(currentPulsar.Neighbours) + 1
+	return nodes - currentPulsar.getMaxTraitorsCount()
+}
+
+func (currentPulsar *Pulsar) handleErrorState(err error) {
+	log.Error(err)
+
+	currentPulsar.clearState()
+}
+
+func (currentPulsar *Pulsar) clearState() {
+	currentPulsar.GeneratedEntropy = [core.EntropySize]byte{}
+	currentPulsar.GeneratedEntropySign = []byte{}
+
+	currentPulsar.CurrentSlotEntropy = core.Entropy{}
+	currentPulsar.CurrentSlotPulseSender = ""
+	currentPulsar.CurrentSlotSenderConfirmations = map[string]core.PulseSenderConfirmation{}
+
+	currentPulsar.ProcessingPulseNumber = 0
+
+	currentPulsar.OwnedBftRow = map[string]*BftCell{}
+	currentPulsar.BftGridLock.Lock()
+	currentPulsar.bftGrid = map[string]map[string]*BftCell{}
+	currentPulsar.BftGridLock.Unlock()
+}
+
+func (currentPulsar *Pulsar) generateNewEntropyAndSign() error {
+	currentPulsar.GeneratedEntropy = currentPulsar.EntropyGenerator.GenerateEntropy()
+	signature, err := signData(currentPulsar.PrivateKey, currentPulsar.GeneratedEntropy)
+	if err != nil {
+		return err
+	}
+	currentPulsar.GeneratedEntropySign = signature
+
+	return nil
+}
+
+func (currentPulsar *Pulsar) preparePayload(body interface{}) (*Payload, error) {
+	sign, err := signData(currentPulsar.PrivateKey, body)
+	if err != nil {
+		return nil, err
+	}
+
+	return &Payload{Body: body, PublicKey: currentPulsar.PublicKeyRaw, Signature: sign}, nil
+}
+
 type ecdsaSignature struct {
 	R, S *big.Int
 }
 
-func CheckPayloadSignature(request *Payload) (bool, error) {
-	return CheckSignature(request.Body, request.PublicKey, request.Signature)
+func checkPayloadSignature(request *Payload) (bool, error) {
+	return checkSignature(request.Body, request.PublicKey, request.Signature)
 }
 
-func CheckSignature(data interface{}, pub string, signature []byte) (bool, error) {
+func checkSignature(data interface{}, pub string, signature []byte) (bool, error) {
 	cborH := &codec.CborHandle{}
 	var b bytes.Buffer
 	enc := codec.NewEncoder(&b, cborH)
@@ -52,11 +119,11 @@ func CheckSignature(data interface{}, pub string, signature []byte) (bool, error
 	var ecdsaP ecdsaSignature
 	rest, err := asn1.Unmarshal(signature, &ecdsaP)
 	if err != nil {
-		return false, errors.Wrap(err, "[ CheckSignature ]")
+		return false, errors.Wrap(err, "[ checkSignature ]")
 	}
 
 	if len(rest) != 0 {
-		return false, errors.New("[ CheckSignature ] len of  rest must be 0")
+		return false, errors.New("[ checkSignature ] len of  rest must be 0")
 	}
 
 	publicKey, err := ecdsahelper.ImportPublicKey(pub)
@@ -67,7 +134,7 @@ func CheckSignature(data interface{}, pub string, signature []byte) (bool, error
 	return ecdsa.Verify(publicKey, b.Bytes(), ecdsaP.R, ecdsaP.S), nil
 }
 
-func SignData(privateKey *ecdsa.PrivateKey, data interface{}) ([]byte, error) {
+func signData(privateKey *ecdsa.PrivateKey, data interface{}) ([]byte, error) {
 	cborH := &codec.CborHandle{}
 	var b bytes.Buffer
 	enc := codec.NewEncoder(&b, cborH)
